beep updates

+++ products/views.py
@@ -1,11 +1,22 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponseForbidden
+
 from .models import Product
 from .forms import ProductForm
 
 def listar_productos(request):
     productos = Product.objects.all()
-    # Retornar los productos al template
-    return render(request, 'products/list_products.html', {'productos': productos})
+
+    # Verificar si el usuario es administrador o cajero
+    es_administrador = request.user.is_admin
+    es_cajero = request.user.is_cashier
+
+    # Retornar los productos al template junto con los roles
+    return render(request, 'products/list_products.html', {
+        'productos': productos,
+        'es_administrador': es_administrador,
+        'es_cajero': es_cajero
+    })
 
 def agregar_producto(request):
     if request.method == 'POST':
@@ -34,7 +45,13 @@ def editar_producto(request, pk):
 
 def eliminar_producto(request, pk):
     producto = Product.objects.get(pk=pk)
+
+    # Verificar si el usuario es administrador
+    if not request.user.is_admin:
+        return HttpResponseForbidden("No tienes permiso para eliminar productos.")
+
     if request.method == 'POST':
         producto.delete()
         return redirect('listar_productos')
+
     return render(request, 'products/delete_product.html', {'producto': producto})
